pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.3.5",
    "opencv-python-headless>=4.11.0.86",
    "pillow>=12.0.0",
    "streamlit>=1.52.1",
    "tensorflow>=2.20.0",
]
